AHA it was rounding

+++ get_bot_response.py
@@ -114,8 +114,8 @@ def get_bot_response(recipient_id, message_text="", attachment = ""):
 
 	else:
 		entity, value, confidence = wit_response(message)
-		print(int(confidence))
-		if int(confidence) > 0.8:
+		print(float(confidence))
+		if float(confidence) > 0.8:
 			print("confident")
 			#response.text  = " : ".join([str(entity), str(value), str(confidence)])
 			response.text = getResponse(entity, value, confidence)
